Fix usernames appearing in all lowercase
For some reason it was getting the usernames from the username_clean
column instead of the username column, this means usernames were
appearing in all lowercase (except for the first character.) This fixes
it so that usernames appear in the correct case.

This is however a problem in already installed installations, because it
means that if this is changed then old usernames that were previously made
lowercase will not match up with new ones. So this sets the default to
make all usernames lowercase so as not to break existing installations.