Fix path concatenation to be consistent
Previously it was required that users enter the PathToPHPBB in their
settings with a / at the end, or else the paths would be concatenated
incorrectly. Worse, the error message concatenated the paths correctly,
such that it would report a missing file that actually existed.

This changes it so that the paths are concatenated consistently, and
also so that it no longer matters whether users add a trailing slash or
not.

Fixes #26.