Simplify config with plugin defaults
Rather than require users to explicitly set all configuration values,
encode the default values in the plugin and allow the array in
LocalSettings.php to override them. This ensures that all values are
defined to known-good and reasonable defaults.

This is entirely backwards-compatible and the defaults in the plugin
were taken from the README.md.

This also reformats the configuration array in README.md to be easier
to read and groups related settings together in a logical way.