Update Auth_phpBBHooks.php
Per https://www.mediawiki.org/wiki/Extension:PluggableAuth the buttonLabelMessage is supposed to be a system message key. With a value of "Log in" you end up with "{Log in}" (with curly braces) on the button since the system can't find that message key. On mediawiki 1.39 the /Special:AllMessages page shows that "login" in the label for the English phrase "Log in".